requirements.txt
flask==3.0.0
flask-restful==0.3.10
python-dotenv==1.0.0
flake8==6.1.0
black==23.11.0
pytest==7.4.3
gunicorn==21.2.0
requests==2.31.0
flask-pymongo==2.3.0
openai==1.60.1
google-generativeai==0.8.4